requirements.txt
astroid==2.2.5
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
python-dotenv==0.10.3
six==1.12.0
SQLAlchemy==1.3.8
typed-ast==1.4.0
waitress==1.3.1
Werkzeug==0.15.6
wrapt==1.11.2
